+++ probe-rs/src/architecture/arm/sequences/nrf.rs
@@ -30,7 +30,11 @@ pub trait Nrf: Sync + Send {
 const ERASEALL: u8 = 0x04;
 const ERASEALLSTATUS: u8 = 0x08;
 
+const APPLICATION_SPU_PERIPH_PERM: u64 = 0x50003800;
+
+const APPLICATION_RESET_PERIPH_ID: u64 = 5;
 const APPLICATION_RESET_S_NETWORK_FORCEOFF_REGISTER: u32 = 0x50005614;
+const APPLICATION_RESET_NS_NETWORK_FORCEOFF_REGISTER: u32 = 0x40005614;
 const RELEASE_FORCEOFF: u32 = 0;
 
 /// Unlocks the core by performing an erase all procedure.
@@ -53,10 +57,20 @@ fn unlock_core(
 
 /// Sets the network core to active running.
 fn set_network_core_running(interface: &mut dyn ArmProbe) -> Result<(), ArmError> {
-    interface.write_32(
-        APPLICATION_RESET_S_NETWORK_FORCEOFF_REGISTER as u64,
-        &[RELEASE_FORCEOFF],
-    )?;
+    // Determine if the RESET peripheral is mapped to secure or non-secure address space.
+    let periph_config_address = APPLICATION_SPU_PERIPH_PERM + 0x4 * APPLICATION_RESET_PERIPH_ID;
+    let periph_config = interface.read_word_32(periph_config_address)?;
+    let is_secure = (periph_config >> 4) & 1 == 1;
+
+    let forceoff_addr = if is_secure {
+        tracing::debug!("RESET peripheral is mapped to secure address space");
+        APPLICATION_RESET_S_NETWORK_FORCEOFF_REGISTER
+    } else {
+        tracing::debug!("RESET peripheral is mapped to non-secure address space");
+        APPLICATION_RESET_NS_NETWORK_FORCEOFF_REGISTER
+    };
+
+    interface.write_32(forceoff_addr as u64, &[RELEASE_FORCEOFF])?;
     Ok(())
 }
 
@@ -109,8 +123,11 @@ impl<T: Nrf> ArmDebugSequence for T {
         }
 
         if self.has_network_core() {
+            tracing::debug!("Setting network core to running");
             set_network_core_running(&mut *interface)?;
         }
+
+        interface.flush()?;
 
         Ok(())
     }